HunarShehar/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-aiplatform==1.36.4
python-dotenv==1.0.0
pydantic==2.4.2
python-multipart==0.0.6 